// foodmine - SVS/frontend/src/app/shared/interfaces/IUserRegister.ts
// Definition of the IUserRegister interface representing user registration details
export interface IUserRegister {
    // Name field for storing the user's full name
    name: string;
  
    // Email field for storing the user's email address
    email: string;
  
    // Password field for storing the user's chosen password
    password: string;
  
    // ConfirmPassword field for confirming the user's chosen password
    confirmPassword: string;
  
    // Address field for storing the user's address details
    address: string;
  }
  